PC-162 Homepage Tweaks 15.01.2024
On the homepage, below the navigation bar, could we please add in a 'sign up now' bar (pdf slide 1) this should replicate the 'sign up now' function in the footer.

On the homepage for each of the callout boxes, please could we add the orange section at the bottom containing the text "Join today and save an average of 10%" (pdf slide 1) this should appear on all hero slide callouts and link to the registration page. Is there also a way that we can move this higher up to ensure that it always sits above the fold on all viewports?

On the homepage the hero slider does not seem to scroll automatically, could this be resolved?

The East arrow on the slider is slightly bigger than the West arrow, could this be resolved?

Could we please re-order the page so that inspiration articles are swapped with cruise packages